add profile-history page
