+++ src/components/app-header/app-header-item/app-header-item.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 
 import itemStyles from './app-header-item.module.css';
-import { Link, useLocation } from 'react-router-dom';
+import { NavLink, useLocation } from 'react-router-dom';
 import { TAppHeaderItemProps } from './app-header-item-types';
 import { TLocationState } from '../app-header-types';
 
@@ -12,7 +12,7 @@ const AppHeaderItem: FC<TAppHeaderItemProps> = ({ text, path = '/', setHover, ch
 
     return (
         <li className={`mt-4 mb-4 pt-4 pb-4 pl-5 pr-5 ${itemStyles.header__item}`}>
-            <Link to={path} className={children ? `mr-2 ${itemStyles.header__iconWrapper}` : ''}>
+            <NavLink to={path} className={children ? `mr-2 ${itemStyles.header__iconWrapper}` : ''}>
                 {children}
                 <p
                     className={`text text_type_main-default ${textColor} ${itemStyles.header__link}`}
@@ -21,7 +21,7 @@ const AppHeaderItem: FC<TAppHeaderItemProps> = ({ text, path = '/', setHover, ch
                 >
                     {text}
                 </p>
-            </Link>
+            </NavLink>
         </li>
     );
 };
