+++ src/components/app/app.tsx
@@ -49,9 +49,7 @@ function App() {
             <AppHeader />
             <main className={styles.main}>
                 <Switch location={modalIngredientOpen || location}>
-                    <Route path="/" exact={true}>
-                        <HomePage />
-                    </Route>
+                    <Route path="/" exact={true} component={HomePage} />
                     <ProtectedRoute path="/login" exact={true}>
                         <LoginPage />
                     </ProtectedRoute>
@@ -67,18 +65,13 @@ function App() {
                     <ProtectedRoute path="/profile" exact={true}>
                         <ProfilePage />
                     </ProtectedRoute>
-                    <Route path="/ingredients/:id" exact={true}>
-                        <Ingredient />
-                    </Route>
-                    <Route path="/feed" exact={true}>
-                        <OrderFeedPage />
-                    </Route>
-                    <Route path="/feed/order" exact={true}>
-                        <OrderPage />
-                    </Route>
-                    <Route>
-                        <NotFoundPage />
-                    </Route>
+                    <ProtectedRoute path="/profile/orders" exact={true}>
+                        <ProfilePage />
+                    </ProtectedRoute>
+                    <Route path="/ingredients/:id" exact={true} component={Ingredient} />
+                    <Route path="/feed" exact={true} component={OrderFeedPage} />
+                    <Route path="/feed/order" exact={true} component={OrderPage} />
+                    <Route component={NotFoundPage} />
                 </Switch>
                 {modalIngredientOpen && (
                     <Route path="/ingredients/:id">
