+++ src/components/feed/feed-order/feed-order.module.css
@@ -1,5 +1,4 @@
 .main {
-    max-width: 584px;
     background: #1C1C21;
     border-radius: 40px;
 }
