+++ src/components/feed/feed.tsx
@@ -6,10 +6,10 @@ import Orders from './orders/orders';
 
 export default function Feed() {
     return (
-        <div style={{ maxHeight: 'calc(100vh - 88px)' }}>
+        <div className={feedStyles.main}>
             <h2 className="text text_type_main-large pt-10">Лента заказов</h2>
-            <div className={feedStyles.main}>
-                <div style={{ maxHeight: 'calc(100vh - 220px)', overflowY: 'scroll' }}>
+            <div className={feedStyles.orders}>
+                <div className={feedStyles.list}>
                     <FeedOrder />
                     <FeedOrder />
                     <FeedOrder />
