+++ src/components/profile/profile-history/profile-history.module.css
@@ -1,0 +1,6 @@
+.main {
+    width: 100%;
+    margin-top: 40px;
+    max-height: calc(100vh - 180px);
+    overflow-y: scroll;
+}
