+++ src/components/profile/profile-history/profile-history.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import styles from './profile-history.module.css';
+import FeedOrder from '../../feed/feed-order/feed-order';
+
+const ProfileHistory = () => {
+    return (
+        <div className={styles.main}>
+            <FeedOrder />
+            <FeedOrder />
+            <FeedOrder />
+            <FeedOrder />
+            <FeedOrder />
+            <FeedOrder />
+            <FeedOrder />
+        </div>
+    );
+};
+
+export default ProfileHistory;
