+++ src/pages/profile.module.css
@@ -1,10 +1,13 @@
 .profile {
-    display: flex;
-    margin-top: 120px;
+    display: grid;
+    grid-template-columns: 1fr 3fr;
+    /*margin-top: 120px;*/
 }
 
 .profile__list {
     list-style-type: none;
+    max-width: 320px;
+    margin-top: 120px;
 }
 
 .profile__item {
@@ -25,46 +28,3 @@
     color: var(--text);
     text-decoration: none;
 }
-
-.profile__inputs {
-    width: 480px;
-    margin-left: 60px;
-}
-
-.profile__inputs > div {
-    width: 100%;
-    padding-bottom: 24px;
-}
-
-.profile__inputs > div > div {
-    width: 100%;
-}
-
-.profile__input_isActive {
-    color: var(--text) !important;
-}
-
-.profile__inputs input {
-    color: var(--text-dark);
-}
-
-/*.profile__input {*/
-/*    color: var(--text);*/
-/*}*/
-
-.profile__info_change {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-}
-
-.profile__info_cancel {
-    color: var(--accent);
-    padding-right: 28px;
-    cursor: pointer;
-}
-
-.profile__info_cancel:hover {
-    color: var(--text);
-}
-
